src/database/connection.py
import mysql.connector
import os
from dotenv import load_dotenv
import logging

load_dotenv()

class DatabaseConnection:
    def __init__(self):
        self.host = os.getenv('DB_HOST')
        self.port = os.getenv('DB_PORT', 3306)
        self.database = os.getenv('DB_NAME')
        self.user = os.getenv('DB_USER')
        self.password = os.getenv('DB_PASSWORD')
        
    def get_connection(self):
        try:
            conn = mysql.connector.connect(
                host=self.host,
                port=self.port,
                database=self.database,
                user=self.user,
                password=self.password,
                autocommit=True
            )
            return conn
        except mysql.connector.Error as e:
            logging.error(f"Database connection error: {e}")
            raise

db_connection = DatabaseConnection()