requirements.txt
# Core Data Engineering & Streamlit
streamlit==1.28.0
streamlit-option-menu==0.3.6
streamlit-extras==0.3.0
streamlit-aggrid==0.3.4.post3
streamlit-authenticator==0.2.1

# Database
mysql-connector-python==8.0.33
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Processing & Analysis
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2
pytz==2023.3

# File Processing & Document Handling
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==0.8.11
docx2txt==0.8
python-pptx==0.6.21
odfpy==1.4.1

# Web Scraping & Crawling
beautifulsoup4==4.12.2
requests==2.31.0
urllib3==1.26.18
lxml==4.9.3
selenium==4.15.2
scrapy==2.11.0
feedparser==6.0.10

# Machine Learning & AI
scikit-learn==1.3.0
joblib==1.3.2
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
gensim==4.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Natural Language Processing
nltk==3.8.1
textstat==0.7.3
wordcloud==1.9.2
contractions==0.1.73
emoji==2.8.0

# Data Visualization
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2
bokeh==3.3.0
altair==5.2.0
chart-studio==1.1.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.1.78
scikit-image==0.21.0
imageio==2.31.6

# Audio Processing
SpeechRecognition==3.10.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Video Processing
moviepy==1.0.3
opencv-python-headless==4.8.1.78

# Archive & Compression
patool==1.12
pyunpack==0.3
zipfile38==0.0.3
rarfile==4.0

# Utilities & Configuration
python-dotenv==1.0.0
python-magic==0.4.27
filetype==1.2.0
mimetypes==0.1.0
pathlib==1.0.1
pathvalidate==3.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3
arrow==1.2.3

# Security & Cryptography
cryptography==41.0.7
pycryptodome==3.19.0
hashlib==20081119

# API & Web Services
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.1
httpx==0.25.2
aiohttp==3.9.1

# Testing & Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Logging & Monitoring
loguru==0.7.2
structlog==23.2.0
sentry-sdk==1.35.0
prometheus-client==0.19.0

# Data Validation & Quality
pydantic==2.5.1
great-expectations==0.18.4
pandera==0.17.1
cerberus==1.3.5

# Workflow & Orchestration
prefect==2.14.5
apache-airflow==2.7.3
luigi==3.3.0

# Distributed Computing
pyspark==3.5.0
dask==2023.11.1
ray==2.8.1

# Cloud Storage
boto3==1.28.65
azure-storage-blob==12.19.0
google-cloud-storage==2.13.0
dropbox==11.36.2

# Big Data Tools
pyarrow==14.0.1
fastparquet==2023.10.1
avro==1.11.3
orc==1.8.2

# Message Queues & Streaming
kafka-python==2.0.2
pika==1.3.2
redis==5.0.1
pulsar-client==3.3.0

# Monitoring & Observability
prometheus-client==0.19.0
grafana-api==1.0.4
elasticsearch==8.11.0

# Containerization & Deployment
docker==6.1.3
kubernetes==28.1.0
celery==5.3.4

# Email & Notifications
smtplib==0.0.1
email-validator==2.1.0
twilio==8.13.0
slack-sdk==3.23.0

# System & OS
psutil==5.9.6
gpustat==1.1
py-cpuinfo==9.0.0

# Internationalization
Babel==2.13.1
googletrans==4.0.0rc1
langdetect==1.0.9

# Other Utilities
tqdm==4.66.1
click==8.1.7
colorama==0.4.6
rich==13.7.0
typer==0.9.0
Jinja2==3.1.2
MarkupSafe==2.1.3